// src/external/api/models/User.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { UserId } from './UserId';
export type User = {
    id?: UserId;
    email?: string;
    createdAt?: string;
    createdBy?: UserId;
    lastEditedAt?: string;
    lastEditedBy?: UserId;
};